requirements.txt
alembic==1.8.1
click==8.1.3
Flask==2.2.2
Flask-Migrate==4.0.0
Flask-SQLAlchemy==3.0.2
gunicorn==20.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.1
psycopg2==2.9.3
psycopg2-binary==2.9.5
PyYAML==6.0
requests==2.28.1
SQLAlchemy==1.4.44
Werkzeug==2.2.2
python-dotenv
